requirements.txt
pytest==7.3.1
selenium==4.9.1
webdriver-manager==4.0.1
pytest-html
pytest-xdist
openpyxl
allure-pytest
pandas
PyAutoGUI
xlwings

